Move packed data source into lenskit-core (#675)
