+++ lenskit-eval/src/main/resources/META-INF/lenskit-eval/methods/packed.properties
@@ -1,1 +1,1 @@
-builder=org.grouplens.lenskit.eval.data.pack.PackedDataSourceBuilder
+builder=org.grouplens.lenskit.data.source.PackedDataSourceBuilder
